pyproject.toml
[project]
name = "pypdfium2-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pypdfium2>=4.30.0",
    "pillow>=10.0.0",
    "fonttools>=4.60.1",
    "pymupdf>=1.26.4",
]
